services/notification_service/pyproject.toml
[project]
name = "notification-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "psycopg2-binary>=2.9.11",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
